ci: run tests on PRs
Previously the checks were only run on branches pushed to the xcp-ng
repo, or, when a branch is pushed to a forked repo, against that fork,
if it has been configured to run GH actions.

This resulted in PRs for branches in a forked repo not to trigger the
test actions, and we would miss any regression while the PR is open,
and only get hit when the regression hits master.

Just activating those actions on pull_request in addition to push
would OTOH have them run twice when we push a PR branch, so we
restrict the runs on push to only happen when pushing to master.  This
has a side-effect of non-PR branches not getting the checks run, which
we may want to fix in the future.

Co-authored-by: Yann Dirson <yann.dirson@vates.tech>
Signed-off-by: Gaëtan Lehmann <gaetan.lehmann@vates.tech>